started working on nonogram structs
split router code out to its own file.
